Move `StorePathWithOutputs` into its own header/file
In the following commits it will become less prevalent.
